requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
sentry-sdk==1.39.1
marshmallow==3.20.1
pytest==7.4.3
mysql-connector==2.2.9
requests==2.31.0
redis==5.0.1
pymongo==4.6.1
